+++ backend/src/common/dtos/create-hospital.dto.ts
@@ -1,5 +1,5 @@
 import { Types } from 'mongoose';
-import { GeoJSONMultiPoint } from '../interfaces/geoJSONMultiPoint';
+import { GeoJSONMultiPoint } from '../interfaces/geoJSONMultiPoint.interface';
 
 export class CreateHospitalDto {
   name: string;
