+++ backend/src/common/dtos/create-preparation.dto.ts
@@ -1,5 +1,5 @@
 import { Types } from 'mongoose';
-import { PreparationType } from '../preparation-type';
+import { PreparationType } from '../preparation.type';
 import { IsDateString } from 'class-validator';
 
 export class CreatePreparationDto {
