+++ backend/src/common/interfaces/geoJSONMultiPoint.interface.ts
@@ -1,0 +1,6 @@
+import { Document } from 'mongoose';
+
+export interface GeoJSONMultiPoint extends Document {
+  readonly type: 'MultiPoint';
+  readonly coordinates: [[number]];
+}
