+++ backend/src/common/interfaces/hospital.interface.ts
@@ -1,6 +1,6 @@
 import { Document } from 'mongoose';
 import { Preparation } from './preparation.interface';
-import { GeoJSONMultiPoint } from './geoJSONMultiPoint';
+import { GeoJSONMultiPoint } from './geoJSONMultiPoint.interface';
 
 export interface Hospital extends Document {
   readonly name: string;
