+++ backend/src/modules/hospitals/hospitals.service.ts
@@ -3,13 +3,16 @@ import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Hospital } from '../../common/interfaces/hospital.interface';
 import { GeoJSONDto } from '../../common/dtos/geojson-point.dto';
-import { PreparationType } from '../../common/preparation-type';
+import { PreparationType } from '../../common/preparation.type';
 import { CreateHospitalDto } from '../../common/dtos/create-hospital.dto';
+import { Preparation } from 'src/common/interfaces/preparation.interface';
 
 @Injectable()
 export class HospitalsService {
   constructor(
     @InjectModel('Hospital') private readonly hospitalModel: Model<Hospital>,
+    @InjectModel('Preparation')
+    private readonly preparationModel: Model<Preparation>,
   ) {}
 
   async create(createHospitalDto: CreateHospitalDto): Promise<Hospital> {
@@ -53,35 +56,7 @@ export class HospitalsService {
       const query = [
         {
           $match: {
-            _id: hospitalId,
-          },
-        },
-        {
-          $project: {
-            preparations: 1,
-          },
-        },
-        {
-          $unwind: {
-            path: '$preparations',
-          },
-        },
-        {
-          $lookup: {
-            from: 'preparations',
-            localField: 'preparations',
-            foreignField: '_id',
-            as: 'preparations',
-          },
-        },
-        {
-          $project: {
-            type: '$preparations.type',
-          },
-        },
-        {
-          $unwind: {
-            path: '$type',
+            hospital: hospitalId,
           },
         },
         {
@@ -90,28 +65,83 @@ export class HospitalsService {
           },
         },
         {
-          $sort: {
-            _id: 1,
-          },
-        },
-        {
           $group: {
-            _id: 0,
-            preparations: {
-              $push: {
-                type: '$_id',
-              },
+            _id: false,
+            all: {
+              $push: '$_id',
             },
           },
         },
-        {
-          $project: {
-            _id: 0,
-            preparations: '$preparations.type',
-          },
-        },
       ];
-      return (await this.hospitalModel.aggregate(query).exec())[0].preparations;
+      return (await this.preparationModel.aggregate(query).exec())[0].all;
     }
   }
+  // async getPreparationsTypes(
+  //   hospitalId: Types.ObjectId,
+  // ): Promise<PreparationType[]> {
+  //   if (Types.ObjectId.isValid(hospitalId)) {
+  //     const query = [
+  //       {
+  //         $match: {
+  //           _id: hospitalId,
+  //         },
+  //       },
+  //       {
+  //         $project: {
+  //           preparations: 1,
+  //         },
+  //       },
+  //       {
+  //         $unwind: {
+  //           path: '$preparations',
+  //         },
+  //       },
+  //       {
+  //         $lookup: {
+  //           from: 'preparations',
+  //           localField: 'preparations',
+  //           foreignField: '_id',
+  //           as: 'preparations',
+  //         },
+  //       },
+  //       {
+  //         $project: {
+  //           type: '$preparations.type',
+  //         },
+  //       },
+  //       {
+  //         $unwind: {
+  //           path: '$type',
+  //         },
+  //       },
+  //       {
+  //         $group: {
+  //           _id: '$type',
+  //         },
+  //       },
+  //       {
+  //         $sort: {
+  //           _id: 1,
+  //         },
+  //       },
+  //       {
+  //         $group: {
+  //           _id: 0,
+  //           preparations: {
+  //             $push: {
+  //               type: '$_id',
+  //             },
+  //           },
+  //         },
+  //       },
+  //       {
+  //         $project: {
+  //           _id: 0,
+  //           preparations: '$preparations.type',
+  //         },
+  //       },
+  //     ];
+  //     return (await this.hospitalModel.aggregate(query).exec())[0].preparations;
+  //   }
+  // }
 }
